Move everything about creating files to `create.hpp`
