+++ src/entry.cpp
@@ -1,5 +1,6 @@
 #include <tmp/entry>
 
+#include "create.hpp"
 #include "utils.hpp"
 
 #include <cstddef>
